requirements.txt
python-telegram-bot[ext]>=21.0
pyyaml>=6.0
mempalace>=3.3.0
